separate CSV table into separate component for reuse elsewhere
